never cache the solver class in plugin_manager
